# openapi.yaml
openapi: "3.0.1"
servers:
  - url: https://chore-management.sralloza.es/api/v1
    description: Production server
info:
  version: "0.1.0"
  title: Chore Management
  contact:
    name: Diego Alloza González
    email: sralloza@gmail.com
    url: https://github.com/sralloza
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

  description: |
    API to manage the chores with the flatmates in a shared flat.

    # Concepts

    - **User**: Each person living in the same flat, a flatmate.
    - **Flat**: A flat. Each user and chore is assigned to a different flat. This API is multi-flat, it can manage the chores of multiple flats. The flat is selected by the `x-flat` header. For more info about the `x-flat` header, please read the [multiflat](#section/Concepts/Multiflat) section.
    - **Chore Type**: A defined chore that needs to be executed weekly in the flat. Each chore type will be assigned weekly to a different flatmate.
    - **Weekly chores**: Chores grouped by week.
    - **Week ID**: Weeks are identified by the year and the week number, like `2022.03` (the third week of 2022).
    - **Ticket**: Number of chores of a specific type that the user is owed. For more info, read the [tickets system](#section/Concepts/Tickets-system) section.

    # Multiflat

    ## Creating a flat

...

    ## Selecting a flat

...

    # Tickets system

...

...

    # Chore Assignements

    You can assign weekly chores using the [createWeeklyChores](#) endpoint.

...

...
paths:
  /flats:
    post:
      tags:
        - Flats
      summary: Create flat
      description: Create new flat.
      operationId: createFlat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlatCreate"
            examples:
              FlatCreate:
                $ref: "#/components/examples/FlatCreate"
      responses:
        "200":
          $ref: "#/components/responses/Flat"
        "400":
          $ref: "#/components/responses/InvalidRequestBody"
        "403":
          description: Invalid create code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                InvalidCode:
                  description: Invalid create code
                  value:
                    message: Invalid create code
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              examples:
                ValidationError:
                  description: Validation error
                  value:
                    errors:
                      - location: body
                        param: name
                        msg: The name field is required

    get:
      tags:
        - Flats
      summary: List flats
      description: List existing flats. This endpoint is not paginated because it isn't expected to manage a large number of flats.
      operationId: listFlats
      security:
        - AdminApiKey: []
      responses:
        "200":
          $ref: "#/components/responses/FlatList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorAdminRequired"
  /flats/{flat_name}:
    get:
      tags:
        - Flats
      summary: Get flat
      description: Get flat details by id
      operationId: getFlat
      security:
        - AdminApiKey: []
      responses:
        "200":
          $ref: "#/components/responses/Flat"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorAdminRequired"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Flats
      summary: Delete flat
      description: Delete flat by id
      operationId: deleteFlat
      security:
        - AdminApiKey: []
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/FlatName"
      responses:
        "204":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorInvalidFlatRequested"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /flats/{flat_name}/settings:
    patch:
      tags:
        - Flats
      summary: Edit flat settings
      description: Edit flat settings
      operationId: editFlatSettings
      security:
        - FlatAdminApiKey: []
      parameters:
        - $ref: "#/components/parameters/FlatName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlatSettings"
            examples:
              FlatSettings:
                $ref: "#/components/examples/FlatSettings"
      responses:
        "200":
          $ref: "#/components/responses/FlatSettings"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                InvalidKeywordMeWithAdminApiKey:
                  $ref: "#/components/examples/InvalidKeywordMeWithAdminApiKey"
                InvalidRequestBody:
                  $ref: "#/components/examples/InvalidRequestBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorInvalidFlatRequested"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /flats/create-code:
    post:
      tags:
        - Flats
      summary: Request a flat-create code
      description: Request a flat-create code, which can be used later to create a flat. The code can only be used to create one flat, after that it will be considered invalid.
      operationId: requestFlatCreateCode
      security:
        - AdminApiKey: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlatCreateCode"
              examples:
                FlatCreateCode:
                  $ref: "#/components/examples/FlatCreateCode"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          # TODO: 401 - Invalid create code
          $ref: "#/components/responses/ForbiddenErrorAdminRequired"
...
  /users:
    post:
      tags:
        - Users
      summary: Register new user
      description: Register a new user
      operationId: createUser
      security:
        - FlatAdminApiKey: []
      parameters:
        - $ref: "#/components/parameters/XFlatHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
            examples:
              UserCreate:
                $ref: "#/components/examples/UserCreate"
      responses:
        "200":
          $ref: "#/components/responses/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorFlatRequired"
        "409":
          $ref: "#/components/responses/Conflict"

    get:
      tags:
        - Users
      summary: List users
      description: List all flat users
      operationId: listUsers
      security:
        - FlatAdminApiKey: []
      parameters:
        - $ref: "#/components/parameters/XFlatHeader"
      responses:
        "200":
          $ref: "#/components/responses/UserList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorFlatRequired"
  /users/{user_id}:
    get:
      tags:
        - Users
      summary: Get user
      description: Get user by id
      operationId: getUser
      security:
        - FlatAdminApiKey: []
      parameters:
        - $ref: "#/components/parameters/UserIDPath"
        - $ref: "#/components/parameters/XFlatHeader"
      responses:
        "200":
          $ref: "#/components/responses/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorFlatRequired"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete user by id
      operationId: deleteUser
      security:
        - FlatAdminApiKey: []
      parameters:
        - $ref: "#/components/parameters/UserIDPath"
      responses:
        "204":
          description: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UnbalancedTickets:
                  description: Cannot delete user with unbalanced tickets
                  value:
                    message: User has unbalanced tickets
                PendingChores:
                  description: Cannot delete user with pending chores
                  value:
                    message: User has pending chores
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenErrorFlatRequired"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  examples:
    AdminAccessRequired:
      description: The admin token is required
      value:
        message: Admin access required
    ConflictError:
      description: The resource already exists
      value:
        message: The resource already exists
...
    Flat:
      value:
        name: flat-example
        api_key: ea427902-2db8-11ed-a261-0242ac120002
        settings:
          assignment_order:
            - 10001
            - 20002
            - 30003
          rotation_sign: positive
    FlatAccessRequired:
      description: An administration flat token is required
      value:
        message: Flat administration access required
    FlatCreate:
      value:
        create_code: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI2MjE2MjA1Mi02ZTg2LTQzNmMtYjg4OS0xMjk1NDM4M2ZiYmIiLCJpc3MiOiJjaG9yZS1tYW5hZ2VtZW50LWFwaSIsImV4cCI6MTY2NTMxNzI3NCwiaWF0IjoxNjY1MzE2OTc0fQ.hbQSg6Asvm7TpAlQDUxnhUlUEqGn2K-DsecFLHDsF3I
        name: flat-example
    FlatCreateCode:
      value:
        code: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI2MjE2MjA1Mi02ZTg2LTQzNmMtYjg4OS0xMjk1NDM4M2ZiYmIiLCJpc3MiOiJjaG9yZS1tYW5hZ2VtZW50LWFwaSIsImV4cCI6MTY2NTMxNzI3NCwiaWF0IjoxNjY1MzE2OTc0fQ.hbQSg6Asvm7TpAlQDUxnhUlUEqGn2K-DsecFLHDsF3I
        expires_at: 2022-10-09T14:07:54.602+02:00
    FlatList:
      value:
        - name: flat-example
          api_key: ea427902-2db8-11ed-a261-0242ac120002
          settings:
            assignment_order:
              - 10001
              - 20002
              - 30003
            rotation_sign: positive
    FlatSettings:
      value:
        assignment_order:
          - 10001
          - 20002
          - 30003
        rotation_sign: positive
    InvalidFlatHeaderUse:
      description: The `x-flat` header was sent without an admin key
      value:
        message: Can't use the x-flat header without the admin key
    InvalidFlatRequested:
      description: The token provided does not grant access to the flat requested
      value:
        message: You don't have permission to access this flat's information
    InvalidKeywordMeWithAdminApiKey:
      description: Can't use the me keyword with the admin API key
      value:
        message: Can't use the me keyword with the admin API key
    InvalidRequestBody:
      description: The JSON sent is invalid
      value:
        message: Request body is not a valid JSON
...
    UnuthorizedError:
      description: Missing API key
      value:
        message: Missing API key
...
    User:
      description: A User
      value:
        username: user-example
        id: 123456789
        api_key: 9ce30976-bb33-4cd7-acb5-b39a7cedb729
    UserCreate:
      description: Register new user
      value:
        username: user-example
        id: 123456789
    UserList:
      description: A list of users
      value:
        - username: user-example-1
          id: 123456789
          api_key: 9ce30976-bb33-4cd7-acb5-b39a7cedb729
        - username: user-example-2
          id: 987654321
          api_key: c26c9865-6d47-4f66-b586-7b489c72624e
    ValidationError:
      description: Validation error
      value:
        message: ValidationError
        errors:
          - msg: Username must be a string
            value: 123
            location: body
            param: username
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      description: Language to use for error messages
      required: false
      schema:
        type: string
        enum:
          - en
          - es
        default: "en"
      example: es
...
    FlatName:
      name: flat_name
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/FlatName"
...
    UserIDPath:
      name: user_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserIDMe"
...
    XFlatHeader:
      name: x-flat
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/XFlatHeader"
  responses:
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            ConflictError:
              $ref: "#/components/examples/ConflictError"
    Flat:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Flat"
          examples:
            Flat:
              $ref: "#/components/examples/Flat"
    FlatList:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FlatList"
          examples:
            FlatList:
              $ref: "#/components/examples/FlatList"
    FlatSettings:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FlatSettingsRequired"
          examples:
            FlatSettings:
              $ref: "#/components/examples/FlatSettings"
    ForbiddenErrorAdminRequired:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            AdminAccessRequired:
              $ref: "#/components/examples/AdminAccessRequired"
    ForbiddenErrorFlatRequired:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            FlatAccessRequired:
              $ref: "#/components/examples/FlatAccessRequired"
            InvalidFlatHeaderUse:
              $ref: "#/components/examples/InvalidFlatHeaderUse"
    ForbiddenErrorInvalidFlatRequested:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            FlatAccessRequired:
              $ref: "#/components/examples/FlatAccessRequired"
            InvalidFlatHeaderUse:
              $ref: "#/components/examples/InvalidFlatHeaderUse"
            InvalidFlatRequested:
              $ref: "#/components/examples/InvalidFlatRequested"
...
    InvalidRequestBody:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            InvalidRequestBody:
              $ref: "#/components/examples/InvalidRequestBody"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            ResourceNotFound:
              value:
                message: Resource not found
...
    Unauthorized:
      description: Missing API Key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            UnuthorizedError:
              $ref: "#/components/examples/UnuthorizedError"
    User:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/User"
          examples:
            User:
              $ref: "#/components/examples/User"
    UserList:
      description: OK
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/User"
          examples:
            UserList:
              $ref: "#/components/examples/UserList"
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"

          examples:
            ValidationError:
              $ref: "#/components/examples/ValidationError"
...
  schemas:
...
    Error:
      description: generic API error
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
          description: A human readable description of what the event represent
      required:
        - message
    Flat:
      description: Registered flat
      type: object
      additionalProperties: false
      properties:
        name:
          $ref: "#/components/schemas/FlatNameString"
        api_key:
          type: string
          description: Admin flat API key
          format: uuid
        settings:
          $ref: "#/components/schemas/FlatSettingsRequired"
      required:
        - name
        - api_key
        - settings
    FlatCreate:
      description: Register new flat
      type: object
      additionalProperties: false
      properties:
        create_code:
          type: string
          format: jwt
          description: Temporal code to create the flat
        name:
          $ref: "#/components/schemas/FlatNameString"
      required:
        - create_code
        - name
    FlatCreateCode:
      description: Code to create a new flat
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          format: jwt
          description: Code to create a new flat
        expires_at:
          type: string
          format: datetime
          description: Expiration date of the code

      required:
        - code
        - expires_at
    FlatList:
      description: List of registered flats
      type: array
      items:
        $ref: "#/components/schemas/Flat"
    FlatName:
      type: string
      description: Name of the flat. This value should be set to the special keyword `me` when using a Flat API key.
      oneOf:
        - $ref: "#/components/schemas/FlatNameString"
        - $ref: "#/components/schemas/FlatNameMe"
    FlatNameString:
      type: string
      pattern: "^[a-z-]+$"
    FlatNameMe:
      type: string
      enum:
        - me
    FlatSettings:
      type: object
      description: Flat settings
      additionalProperties: false
      properties:
        rotation_sign:
          type: string
          enum: [positive, negative]
          default: positive
          description: When assigning weekly chores, should the rotation be positive or negative
        assignment_order:
          type: array
          description: Order in which users are assigned chores
          items:
            $ref: "#/components/schemas/UserID"
    FlatSettingsRequired:
      allOf:
        - $ref: "#/components/schemas/FlatSettings"
        - required:
            - rotation_sign
            - assignment_order
    SingleValidationError:
      description: Validation error
      type: object
      additionalProperties: false
      properties:
        msg:
          type: string
          description: A human readable description of what the validation error
        value:
          description: The value that caused the validation error
        param:
          description: Name of the param that caused the validation error
        location:
          type: string
          enum: [body, cookies, headers, params, query]
          description: Path to the attribute that caused the validation error
      required:
        - msg
        - location
        - param
...
    User:
      description: Registered user
      type: object
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/UserCreate"
        - properties:
            api_key:
              type: string
              format: uuid
              description: User's API token
          required:
            - api_key
    UserCreate:
      description: Register new user
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/UserID"
        username:
          type: string
          description: User's nickname
          maxLength: 25
          minLength: 2
      required:
        - id
        - username
    UserIDMe:
      description: User ID. This value should be set to the special keyword `me` when using a user API key.
      anyOf:
        - type: string
          enum: [me]
        - $ref: "#/components/schemas/UserID"
    UserID:
      description: User identifier
      type: integer
      format: int64
    ValidationError:
      description: Validation error
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/SingleValidationError"
    XFlatHeader:
      type: string
      description: Name of the flat. For more info please refer to the [multiflat](#section/Concepts/Multiflat) section
...
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: x-token
    FlatAdminApiKey:
      type: apiKey
      in: header
      name: x-token
    UserApiKey:
      type: apiKey
      in: header
      name: x-token